% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shist.R
\name{choose_bin_width}
\alias{choose_bin_width}
\title{choose_bin_width}
\usage{
choose_bin_width(x)
}
\arguments{
\item{x}{a vector of values to derived a step size from.}
}
\description{
A helper funciton for shist() that chooses the slider step when none is provided.
The idea is to choose a whole number that represents a represents a change in 5 percetiles around the median of the distribution.
The distribution is hopefully changing slowly around the median and thus the interval will be the an estimate of the smallest useful delta in bin size.
}

