% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shist.R
\name{shist}
\alias{shist}
\title{shist}
\usage{
shist(x, bin_step = NA)
}
\arguments{
\item{x}{a vector of values to plot in a histogram.}

\item{bin_step}{an optional number that defines the size of the increments on the bin width slider.
If left unspecified a whole number step size is derived from the data using an unspecified algorithm.}
}
\description{
shist takes a vector of values and plots a ggvis histogram with a slider to select the bin width.
This saves you having to re-plot over and over trying to find the bin width that best suits the inference you want to make.
Once the slider has been clicked on you can use the left and right arrow keys to nudge the slider and ggvis handles updating the histogram
in real time.
}

